Creating more unit tests, moving others to correct package
