+++ sample/src/test/java/com/vimeo/sample/model/AbstractDataListTest.java
@@ -1,0 +1,17 @@
+package com.vimeo.sample.model;
+
+import com.vimeo.sample.Utils;
+
+import org.junit.Test;
+
+/**
+ * Created by restainoa on 2/2/17.
+ */
+public class AbstractDataListTest {
+
+    @Test
+    public void typeAdapterWasNotGenerated() throws Exception {
+        Utils.verifyNoTypeAdapterGeneration(AbstractDataList.class);
+    }
+
+}
